Move Ably sandbox code to test/lib
I’ll reuse it in the upcoming integration tests.
